requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cohere==2.8.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
numpy==1.23.4
pandas==1.5.1
psycopg==3.1.4
psycopg-binary==3.1.4
pydub==0.25.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
SpeechRecognition==3.8.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
